Refactor __main__ into smaller progs/ files
